+++ include/docker/docker-ce_debain.sh
@@ -4,45 +4,28 @@
 #
 # Github: https://github.com/zhwknight/setup-scripts
 
-# Remove old version
-sudo apt remove -y \
-  docker.io \
-  docker-doc \
-  docker-compose \
-  podman-docker \
-  containerd \
-  runc \
-  ;
+# Also convenient to use the following script to install docker
+# curl -fsSL https://get.docker.com -o get-docker.sh
+# sudo sh ./get-docker.sh --dry-run
 
-# Install packages to allow apt to use a repository over HTTPS
-sudo apt install -y \
-  ca-certificates \
-  curl \
-  gnupg \
-  lsb-release \
-  ;
+# Uninstall all conflicting packages
+for pkg in docker.io docker-doc docker-compose podman-docker containerd runc; do sudo apt-get remove $pkg; done
 
-# Add Docker’s official GPG key and set up the stable repository
+# Add Docker's official GPG key:
+sudo apt-get update
+sudo apt-get install ca-certificates curl
 sudo install -m 0755 -d /etc/apt/keyrings
-curl -fsSL https://download.docker.com/linux/debian/gpg -o /tmp/docker.asc
-mv /tmp/docker.asc /etc/apt/keyrings/docker.asc
+sudo curl -fsSL https://download.docker.com/linux/debian/gpg -o /etc/apt/keyrings/docker.asc
 sudo chmod a+r /etc/apt/keyrings/docker.asc
+
+# Add the repository to Apt sources:
 echo \
   "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/debian \
   $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
   sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+sudo apt-get update
 
-# Update the apt package index
-sudo apt update
-
-# Install the latest version of Docker Engine - Community and containerd, or go to the next step to install a specific version
-sudo apt install -y \
-  docker-ce \
-  docker-ce-cli \
-  containerd.io \
-  docker-buildx-plugin \
-  docker-compose-plugin \
-  ;
+sudo apt-get install docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
 
 # Adding your user to the "docker" group
 sudo groupadd docker